Remove visit_id and add claim_id and line_id where applicable
Since we have now removed visits and encounters we need to remove visit_id from tables and replace with claim_id and/or line_id.

Also removed clinical_code_id from clinical_codes table which was original there to link a diagnosis with a procedure but this is now handled by the lines table.